requirements.txt
cffi==1.17.1
cryptography==44.0.0
MouseInfo==0.1.3
pillow==11.0.0
pycparser==2.22
PyGetWindow==0.0.9
PyMsgBox==1.0.9
pyotp==2.9.0
pyperclip==1.9.0
PyQt5==5.15.11
PyQt5-Qt5==5.15.16
PyQt5_sip==12.16.1
PyRect==0.2.0
PyScreeze==1.0.1
python3-xlib==0.15
pytweening==1.2.0
pyzbar==0.1.9
qrcode==8.0